refactor: extract ArtistAccessSelector content into a dedicated component
- Moved the inline access management logic from ArtistAccessSelector to a new ArtistAccessSelectorContent component for improved code organization and readability.
- This change simplifies the ArtistAccessSelector component and enhances maintainability by separating concerns.
